+++ src/Phan/Language/Element/Func.php
@@ -282,6 +282,7 @@ class Func extends AddressableElement implements FunctionInterface
 
     /**
      * @return \Generator
+     * @phan-return \Generator<Func>
      * The set of all alternates to this function
      */
     public function alternateGenerator(CodeBase $code_base) : \Generator