+++ src/Phan/Language/Element/Method.php
@@ -498,6 +498,7 @@ class Method extends ClassElement implements FunctionInterface
 
     /**
      * @return \Generator
+     * @phan-return \Generator<Method>
      * The set of all alternates to this method
      */
     public function alternateGenerator(CodeBase $code_base) : \Generator